[Frontend] app pages split to page elements
