+++ Mobil/src/screens/CarbonFootprint/index.ts
@@ -1,0 +1,2 @@
+export { CarbonFootprintScreen } from './CarbonFootprintScreen';
+export * from './types';
