+++ Mobil/src/screens/CarbonFootprint/types.ts
@@ -1,0 +1,5 @@
+import { Theme } from "../../utils/theme";
+
+export interface CarbonFootprintScreenProps {
+  theme: Theme;
+}
