+++ Mobil/src/screens/Disease/index.ts
@@ -1,0 +1,4 @@
+export { DiseaseScreen } from './DiseaseScreen';
+export { CameraView, canUseCameraComponent } from './CameraView';
+export { PhotoPreview } from './PhotoPreview';
+export * from './types';
