+++ Mobil/src/screens/Home/CameraController.tsx
@@ -1,0 +1,20 @@
+import { useEffect } from "react";
+import { useThree } from "@react-three/fiber/native";
+import { CameraConfig } from "./types";
+
+interface CameraControllerProps {
+  config: CameraConfig;
+}
+
+export function CameraController({ config }: CameraControllerProps) {
+  const { camera, invalidate } = useThree();
+
+  useEffect(() => {
+    camera.position.set(...config.position);
+    (camera as any).fov = config.fov;
+    (camera as any).updateProjectionMatrix();
+    invalidate();
+  }, [config, camera, invalidate]);
+
+  return null;
+}
