+++ Mobil/src/screens/Login/types.ts
@@ -1,0 +1,5 @@
+export interface LoginScreenProps {
+  theme: any;
+  onLoginSuccess: () => void;
+  onSkip: () => void;
+}
