+++ Mobil/src/screens/Settings/index.ts
@@ -1,0 +1,2 @@
+export { SettingsScreen } from './SettingsScreen';
+export * from './types';
