+++ Mobil/src/screens/Settings/types.ts
@@ -1,0 +1,17 @@
+import { Theme } from "../../types";
+
+export type ThemeMode = 'light' | 'dark' | 'system' | 'weather';
+
+export interface SettingsScreenProps {
+  theme: Theme;
+  isDark: boolean;
+  themeMode: ThemeMode;
+  onThemeModeChange: (mode: ThemeMode) => void;
+  onLogout: () => void;
+}
+
+export interface ThemeOption {
+  mode: ThemeMode;
+  label: string;
+  icon: string;
+}
