+++ Mobil/src/screens/index.ts
@@ -1,6 +1,6 @@
-export { LoginScreen } from './LoginScreen';
-export { HomeScreen } from './HomeScreen';
-export { DiseaseScreen } from './DiseaseScreen';
-export { TimetableScreen } from './TimetableScreen';
-export { SettingsScreen, type ThemeMode } from './SettingsScreen';
-export { CarbonFootprintScreen } from './CarbonFootprintScreen';
+export { LoginScreen } from './Login';
+export { HomeScreen } from './Home';
+export { DiseaseScreen } from './Disease';
+export { TimetableScreen } from './Timetable';
+export { SettingsScreen, type ThemeMode } from './Settings';
+export { CarbonFootprintScreen } from './CarbonFootprint';
